gui: wizard: split wizard and renamed to WelcomeWizard
